part-3-frontend/requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.5.0
python-dotenv==1.0.0
Werkzeug==2.3.4
WTForms==3.0.1
